Update package name and fix mojo properties
